Add textfield to resource chooser for @string resources
This changeset adds a new textfield to the bottom of the resource
chooser in some scenarios, such as when editing @string resources via
Edit Text in the layout editor.

This text field updates to show the @string/resourceName value
whenever you update the selection in the filter, but it also allows
you to edit the text to anything else, such as a literla String (not a
resource). This makes it easy to edit simple labels. See the following
adt-dev thread for some background on this:

   http://groups.google.com/group/adt-dev/browse_thread/
   thread/d53e71c2970ce369?pli=1

There is also a new label below the text field which shows the
"resolved value". For example, if you've chosen "hello" in the filter
list, the new value text will show "@string/hello", and the new label
will show "Hello World". This will hopefully make it more obvious what
it is you're editing. And, similar to the widget render preview for
drawable resources, this makes it easier to browse and pick labels,
especially when browsing framework resources. This is currently only
enabled for @string and @dimen resources, but we could conceivably add
it for other resource types too.

Longer term, we need inline editing of widgets (where values entered
automatically get translated into resources), but this is a stopgap
measure which makes it easier to build simple apps quickly, and which
still makes it simple to create new strings or select existing ones
from the filter list. (And we have both Lint and the Extract String
refactoring to help identify and fix hardcoded strings later on.)

While working on this I also discovered the root cause for this bug:
  20589: The reference chooser is not usable for framework styles
Turns out org.eclipse.ui.dialogs.AbstractElementListSelectionDialog
has a handleEmptyList() method which calls setEnabled(false) on the
filter list when the list is empty. Crucially, there is no
corresponding setEnabled(true) call when the list is made non-empty!
When the user switches between framework and project resources, the
list is temporarily empty, so the list gets disabled, and that's why
it's not responding to menu clicks at all. This changeset also fixes
that bug.

Change-Id: Id441104d617fa4440daba3da2dc63e2c999f48c1