Add "Extra Text" lint detector.
This new lint check looks for XML text content in resource files where
text is not expected, such as layout files. This is typically the
result of a typo.

This CL also improves the command line XML parser's position handling
such that it can handle positions on text nodes (until now it just
supported element and attribute positions), and similarly for Eclipse.

Change-Id: I6f69a8dc725a5b5b19f8ad5bb889ae25ee82600a