Hide the animated-rotate metadata from the drawable descriptors
This changeset removes the metadata for <animated-rotate>
drawables. Turns out the underlying class,
android.graphics.drawable.AnimatedRotateDrawable, is marked
hidden. Therefore, the <animated-rotate> and its associated attributes
should not be shown to the user since they cannot be used. This is the
subject of issue http://b.android.com/19248 .

While looking at this I added some instrumentation to the addElement
method (which adds the manually registered style classes and their
attributes). This validates that all added attributes are referenced
in the public.xml file. The above <animated-rotate> class is the only
class that was not public. However, that's only among the manually
registered metadata for animators, drawables and color lists.  For the
overall attributes (on layouts etc), there are many attributes picked
up from the attrs.xml file that do not appear in public.xml.  In a
subsequent CL we need to add filtering to that import phase as well.

Change-Id: Ia6fba23ca47b12cc9a13cc7e203b6d5849dfe103