Fix locale handling of uppercase/lowercase
This changeset fixes issue
23747: i and İ character problem in turkish operating Systems...
and probably many other bugs in the Turkish locale.

Basically, we had a lot of String.toLowerCase() and
String.toUpperCase() calls. This performs locale sensitive
conversions, which in many cases is NOT what we want; for "machine
readable" conversions we should be using Locale.US which performs no
special cases.

For more, see
http://developer.android.com/reference/java/util/Locale.html#default_locale

Change-Id: I996b0e70fb377e8dae484c5811deb8bc9afb684c