ADT: single base class for XML editors.
This re-architecture the way the various XML editors are
handled by ADT.

Before, we had various classes all deriving from AndroidXmlEditor,
all listed in the plugin.xml as handled the "xml" extension. The
right one was set by AdtPlugin when monitoring changed resources.

The new structure is to have a single AndroidXmlCommonEditor.
This derives from AndroidXmlEditor right now, but in phase 2 will
merge them back in a single class.

Each XML editor implements an XmlEditDelegate. The
delegator iterates through the delegates to find the correct
one that can handle a given file and then uses it.

After this CL, we'll need a few extra passes:
- Understand what all editors delegates are doing
  and merge it in the base class. There's currently a bit
  of code duplicationg we could avoid.
- Change the ManifestEditor to the new structure.
- Change LayoutEditor to the new structure (already prepared
  but in a different CL)

Change-Id: I6cb045a8fa39386fcc0ce854b45fa8580fae411e