SDK Manager: fix missing label in Source categories.
When listing packages by "repository" (e.g. by source URL),
the name of external source isn't always set. It turns out
that when processing packages, we load local packages first
and only their source URL is known, not the "UI name" of the
source so this results in categories with no known UI name
*if* they have at least one local package installed.

This changes the package loader so that it will try to adjust
the label of a source even if it's loaded after the category
has been created.

Change-Id: I285fdefdae99c635d9fd139ca50d17706eea3ae6