gltrace: map calls to frame after sort by start time.
If the trace contains multiple contexts, calls from multiple
contexts may be out of order. Hence the calls are sorted by start
time when the trace is read.

Currently, calls are assigned to frames when they are first read
in from the trace file. This patch performs this assignment once
the calls are sorted.

Change-Id: I7a1b51fac18249cb965756b567d4245d783647ee