gltrace: do not keep the protobuf message in memory
Currently, the original protobuf message is retained in
memory as part of a GLCall. It turns out that this takes
up a significant amount of memory.

This patch extracts the necessary fields out of the
protobuf and puts them in GLCall. No references to the
protobuf are maintained in memory anymore, although
they can be recreated by reading in the file again.

Change-Id: I93cb8a41a1d8b274ea14b9ae4cde034f0ce70f83