gltrace: state view improvements
The GL State View currently utilizes a single tree and switches
the underlying model depending on which editor is active. This causes
loss of context when switching between editors.

This patch makes the state view behave more like the Eclipse Outline
View. State View now inherits from PageBookView, and it allows each
editor to supply whatever controls it wants to display in the view.
Each editor supplies an instance of StateViewPage based on the state
of the trace in the editor.

This patch also moves the state information out of the GLTrace object
and into the StateViewPage.

Change-Id: Ied827ffc53ebba835f1617126068eb3de50f12d4