Split ProGuard file into two halves
This changeset splits the proguard.cfg into two halves:

(1) All the general Android settings go into
    $ANDROID_SDK/proguard/proguard-android.txt.

    This defines shrinking rules like keep custom views, etc. The
    crucial point is that this information is maintained and updated
    by Tools updates, so whenever new APIs are added to Android, or
    whenever bugs are found in the configuration such as flags needed
    to work with Dalvik, we can make the updates - we don't have old
    snapshots living on in projects.

(2) Any project specific settings go to proguard-project.txt in the
    project.

(3) The proguard.config property in project.properties now refers to a
    *path* of configuration files, which are all passed to ProGuard in
    the given order. The code which processes this setting will
    substitute android.sdk.home and user.home variables, so the path
    does not have to be hardcoded to point to the project-android.txt
    file.

    The default project templates have been updated to include a
    commented out configuration setting up proguard as described
    above.

The default proguard file name was changed from proguard.cfg to
proguard-project.txt such that it can be directly opened in Eclipse
and to make it clear it's an editable text file.

Lint was updated to find the Proguard file via the proguard.config
property as well as via the old and new default names for projects not
enabled with ProGuard.

A subsequent CL will add a lint check which identifies projects
containing the old setup (full local configuration) and offer to
replace it with the new setup.

Change-Id: I44b4c97a160114c2382f02f843c95486a0dc9d6b