Prevent cycles in layout editor (and detect with lint)
If you specify <style name="Foo" parent="@style/Foo.Bar"/> then the
layout editor hangs with infinite recursion because as it walks up the
parent chains, it first goes from Foo.Bar (thanks to the style
definition), and from there it goes to Foo (because Foo is the implied
parent of Foo.Bar) and so on.

This changeset adds a visited set to the isTheme() method to prevent
boundless recursion in the presence of cycles.

It also adds a check for this condition to the lint style cycle
detector.

This fixes
http://code.google.com/p/android/issues/detail?id=24385
StackOverflowError when opening any layout xml file in Eclipse

Change-Id: I5c52e77ab7e81ea528f3b8fcdd052bff9447b2c9