Add quick fix for the UseCompoundDrawable lint warning
This changeset adds a quickfix for the "Use Compound Drawable" lint
warning. That warning identifies a LinearLayout which contains exactly
one TextView and one ImageView (in either order, and the linear layout
can be in either orientation), and suggests replacing it by just the
single TextView along with a drawable attribute. The description alone
wasn't clear, so the quickfix performs the conversion automatically.

In addition to the conversion it also needs to transfer any layout
param references from the old LinearLayout to the TextView (and assign
it an id if necessary), so the quickfix is implemented as a visual
refactoring. It's also made available from the Visual Refactoring
menu.

This changeset also updates a couple of unrelated golden files in the
refactoring unit tests to track a message change a few months back.

Change-Id: Iecb544d196fdd4bcabf13a3fdb82ef1ea063f4c2