Add render logger hyperlinks to define layout_width and height
If a layout requires the layout_width or layout_height to be defined
(which all layouts except GridLayout do), and an XML layout does not
define it, then the view will throw an exception during render, which
states that the attribute must be defined, but does not state which
element is missing it.

This changeset adds recognition code in the RenderLogger which detects
when this render failure happens, and when listing the render errors
it performs a check in the XML to identify the offending element(s)
and lists them. It also adds hyperlinks to directly set the value to
either wrap_content or fill_parent (or match_parent if the API level
>= 8).

This addresses 24624: ADT GLE tells me to supply layout_width/height
attributes... but to what?

Change-Id: I967a72e61dc7d494c213111a86bd5865943e07fe