gltrace: add a texture view to display the texture image
This patch adds a new view that displays the texture image
in an ImageCanvas.

The code shared between the FrameBufferView and the TextureView
have now been moved to an abstract ImageViewPart class.

The GL state view provides selection events that are listened
to by the texture view. When a texture property is selected,
the appropriate texture image is displayed in the texture view.

Change-Id: I5f192c88ace447e267d1b3c110ab9bfd39864321