Build Monitor RCP as part of SDK (Mac, Linux and Win)
This adds a new SDK/tools/monitor[.bat] wrapper
as well as SDK/tools/lib/monitor-x86[_64]/ directories
that contain the actual RCP application.

The monitor shell wrappers + bat have not been properly tested yet.

The goal of this CL is just to build the RCP. In a following
CL we'll adjust the wrapper and the actual binary location
in the SDK.

Cross-dependency: Requires build.git change Ifb31e2e3.

Change-Id: I7afbd63badfb81fd5d3d899de672e02525a49bc2