gltrace: color glDraw() functions differently
This patch makes the glDraw*() calls stand out in the table and
in the duration minimap by coloring them differently than the rest
of the functions.

This patch also changes displays the contents of the framebuffer
in the framebuffer view as soon as a frame is selected (as
opposed to updating the view only when the eglSwapBuffer function
is selected).

Change-Id: I749e277e468003612895c45d90a5874b2a84e4d3