Added check for mText property in getText()
Sometimes the node's property is not 'text:mText' but just 'mText'.
A new test was added to getText() to check for both properties and not
throwing unnecessarily the RuntimeException("No text property on node").

This has been tested by monkyrunner tests using EasyMonkeyDevice that runs
correctly after this patch is applied.

Change-Id: I5399169b906e28eeee0a17ab298a724cc5bb9b51
Signed-off-by: Diego Torres Milano <dtmilano@gmail.com>