Add support for exporting one HTML report for each linted project
If you point the --html flag to a directory rather than a file (or to
a nonexistent file which does not have a file extension, such as
/tmp/lintoutput/) then lint will create a separate report for each
linted project and write each one into the export directory.

It also writes the project name into the report title.

(This changeset also performs some other cleanup to make delegating
HTML reporting easier now that it's done by the multi-project
exporter.)

Change-Id: I460e9fad5ae7176d066550f6193ba79a7668e416