Add @SuppressLint support for Java bytecode detectors
This changeset adds support for suppressing lint warnings by
annotatating classes and methods. (Finer granularity will be available
for Java source detectors, but does not work for byte code detectors
since those annotations disappear from the .class file).

It also rewrites two of the existing bytecode detectors to use the ASM
DOM instead of the old visitor pattern such that it can pass a
MethodNode to the Lint driver to check for suppress annotations.
While doing that I also make the field getter detector print the name
of the field that the getter call can be replaced with.

Change-Id: I8f7accb0219e6e9b3c499f16e7c0eb46853bbe04