Write lint overview report for multi-report
If lint writes out multiple HTML reports (e.g. when specifying a directory to the --html flag), and
there are more projects than just one, write an overview doc which contains a table of all the
projects and the corresponding error and warning counts in decreasing order, with links to
the individual reports.

Also ensure that there are no case-insensitive filename clashes.

Change-Id: I4dcda14a625863e71022734088a027fc8ac357d0