gltrace: Display Calls in a Hierarchy
This patch uses the gl extensions glPushMarker and glPopMarker in
the trace file to construct call hierarchies. The call hierarchies
are then displayed in a tree view.

In order to perform this, the main widget in the GLFunctionTraceViewer
has been changed from a Table to a Tree. The model for the tree
is GLCallNode, which is a thin wrapper on top of a GLCall.

A hierarchy can only be displayed if there is a single context
that is being viewed. If a trace file has multiple contexts, then
a context selector Combo box is provided that controls which
context is being viewed.

Change-Id: I043db99c15dfcc179fcffb656f3ffd2d9864f848