Add marker resolution for adding @SuppressLint to Java files
This changeset adds a new marker resolution for adding @SuppressLint
annotation (or modifying an existing @SuppressLint) in the method,
field or class surrounding a lint warning. It searches up in scope
and adds one suggestion for each enclosing applicable scope.

This changeset also makes the suppress id's be case insensitive such
that a codestyle similar to the one used by @SuppressWarnings is
possible.

Change-Id: I0a6a80c4e2d526f1c8fc6aedd95674a25e6a22f0