Include suppress help in HTML report, and tweak CSS
First, include the suppress help as a section in the bottom of the
HTML report, and link to it from each issue explanation.

Second, tweak the generated HTML and CSS a bit: Fix <a> element
generation to be properly closed (which avoids issue where the first
link in each location list did not have the right style), make links
in the More Info have a blue color to stand out more, and add a Holo
like gradient in the background color for browsers that support it.

Change-Id: I3078cb1f1a5112dc6360816c4a6a6c37771ab6fc