Add lint rule checking for showAsAction="always" in menus
It's better to use "ifRoom" instead. The check allows a menu to have a
couple of "always" items if there is also at least one "ifRoom" item
in the menu. In Java code, the lint check will complain if it finds
one or more references to MenuItem.SHOW_AS_ACTION_ALWAYS but no
references to SHOW_AS_ACTION_IF_ROOM.

Change-Id: Ib13f89920bd3919ffd7fc0f0137e90103e8850c3