ddms: fix deadlock
A deadlock can occur between:
    - a thread calling handleChunk() resulting in Device.getClients()
    - a thread calling processIncomingJDWPData() resulting in
      dropClient()

For a full trace resulting in deadlock, see attachments in bug 24926.

This patch fixes the deadlock by moving the entire synchronization
into MonitorThread. Originally, processIncomingJdwpData, performs
locking as follows:
    synchronized(monitorThread) {
        synchronized(device.getClientList() {
            synchronized(monitorThread.mClientList) {
                synchronized(device.getClientList() {

This patch changes the structure of locking to look like:
    synchronized(device.getClientList()) {
    }

    synchronized(monitorThread) {
        synchronized(monitorThread.mClientList) {
            synchronized(device.getClientList() {

As a result, device.getClientList() is always locked *after*
monitorThread.mClientList. This matches the order of locking
performed with handleChunk(), and should solve this deadlock
scenario.

Change-Id: Ie7a9efa897568af4450f0e23b2175ee66e127d7d