gltrace: Use half-open ranges when defining frames.
A GLFrame maintains the start and end indices of the GL Calls
composing that frame. Currently, the indices comprise a closed
range (i.e. include both the call at the start index and the call
at the end index). This patch makes it a half open range (the
end index is excluded) to make things easy to follow.

Change-Id: Ie44b56a32d002595493b11f496ebf123072f8962