gltrace: Add a low cost way to maintain per call properties.
For most GL Calls, we don't need to save anything but its string
representation from its protocol buffer. But for a number of calls,
additional data is required. But we don't want to save the entire
protobuf (due to its high memory consumption).

So this patch adds a generic properties holder to the GLCall
structure. While the trace file is parsed, certain calls may
save data from the protobuf in this structure and these can be
retrieved at runtime.

This patch also utilizes this feature to save the marker name
for every glPushGroupMarker call. Only these markers are displayed
in the trace view as opposed to the full gl call.

Change-Id: If989b3a35704fd0017a944e58589d2122a9a0334