gltrace: use DetailsView to display state details.
This CL replaces the texture view with a generic details view
that can be used to show the details for any state property.

The details view is structured as a page book view that has
a stack layout capable of showing different controls. It has
a list of IStateDetailsProviders to which it delegates the
task of providing details for a given GL property.

Currently, there are 2 such details providers - one for
texture images and one for shader source.

Change-Id: I1ead8848116b0a5889d272f69b49610590937a8f