SDK Repository: remove <extra> from repository.xml
To simplify and avoid XSD duplication, remove extra
packages from the repository.xsd. Instead they should
be defined via an addon.xml.

Note that -repository-6.xsd is prefixed with a dash,
meaning the SDK Manager 17 knows about it and can enforce
it but the build server will continue creating repos
in v5 for now. Once r17 is deployed, we can make the
xsd v6 mandatory for r18.

Change-Id: I00d4644378bcc95bfd9ceb7eede0c8c720926f74