Add namespace-typo detector
This changeset adds a new lint detector which looks for typos in the
namespace declaration for Android. If it sees the prefix "android" it
ensures that the URI matches exactly (and case-sensitively)
   http://schemas.android.com/apk/res/android

If the prefix is "a", it also checks for the above prefix, but only
warns if the difference is "close" (edit distance <= 3) such that
it belives it's a typo, not some unrelated prefix bound to "a".

(This was requested in http://b.android.com/25449)

Change-Id: Ic6dee77a300fc048cb9bee3e455f533b858ef6bb