Add suppress-this to Lint View, and preserve tree expansion state
This changeset adds a new toolbar button to the Lint View: "Suppress
this error with an annotation/attribute". This is enabled when the
selected warning is for a Java or XML file, and it adds the annotation
or suppress attribute in the relevant source file for the warning.

It also adds some handling to preserve the expanded tree state in the
Lint View when the view is refreshed, which happens automatically
whenever a fix is applied or a file is saved and the marker set
changes.

Change-Id: I393ee5f6e062a0e08a7e9457a03f37d8899a4df8