Add onClick check for lint
This lint check makes sure that if a layout references
android:onClick="foo" then one of the Java classes provide a
        public void foo(View)
method. It also looks for simple typos, and warns if the access
modifiers are not right.

(It also moves some ASM-related utility code out of specific detectors
and into the ClassContext class)

Change-Id: Ifb0820221175a0760b83992f54e99f761b4d7097