Lint rule to catch edit-usage of TextViews (should be EditText)
TextView provides a number of attributes related to text editing.
However, these are generally intended for the EditText subclass of
EditText. This lint check looks for <TextView> elements in layouts
where one or more of the edit-related attributes (such as inputType)
are defined.

It also updates the deprecation detector to check various deprecated
attributes (it only checked for deprecated elements before this.)

Change-Id: I1b94902d906cd0bee6a6564368125a68b3c9bb2a