Use lint error mechanism for config issues like no classes found
If Lint cannot find the .class files for a project, it cannot run any
of the class-file based checks (such as the NewApi check).

This changeset adds a new category and issue id, "Lint Error", for
these types of issues. In HTML reports, these errors are listed at the
top. The issue explanation states that these errors don't represent
bugs in the user's code, but that lint was not able to check certain
things for the reasons given.

In the case of no .class files found, it asks whether the project
needs to be built first.  It also uses these lint errors to emit
errors in processing lint.xml configuration files.

(Note that if you don't want to see these types of errors, you can
suppress it via --disable LintError.)

Change-Id: Ifc2f55566f3a0cd20189d43e4205201bc21ee280