Fix lint library project configuration handling
A lint.xml file in a library project should be able to suppress lint
errors in that project. This changeset fixes a couple of bugs which
prevented this from working: In Eclipse, the configuration was cached,
which meant a single lint run through multiple projects would just use
the configuration of the first encountered project.  And from the
command line driver, the code to handle --check handling would
override configuration severity in one scenario.

Change-Id: Ib08e2b14582ab53e5ab89244738d1bd77839610e