ADT fix save on layout configuration change.
When a layout is opened and its configuration is changed,
the new delegate layout wasn't properly saving the layout
before changing files.

This is due to another instance of the delegate code calling
a method on the delegate whereas it should have called a method
with the same exact name on the editor.
To reduce this confusion, this CL adds a 'delegate' prefix to
methods which are common in the CommonXmlEditor and the delegates.
At the expense of being a bit more verbose it makes it clearer
which method is being invoked.

Change-Id: I34e620a9d25272fe75dab54279161b04a961a598