Always use dynamic IDs for unknown project resources.
In the past we've only used dynamic IDs for missing resources of
type ResourceType.ID because they can be created by layout files that
are not saved.

However if auto-refresh and auto-build are both disabled, there can
be operations that create resources that won't be present in R.java yet
but are still needed for rendering. So we need to create dynamic
IDs for them too.

This commit reuse the new DynamicIdMap from the layoutlib since we needed
one that included resource type instead of just resource name.
The dynamic map is reset every time the resources are compiled and we get
a new set of ID from aapt.

Change-Id: I059c93f63f32a2e389783d21c66c7d8c7d818a2d