Handle nulls from Element.getAttribute (fix for issue 25668)
Element.getAttribute isn't supposed to return null, but in Eclipse it
sometimes does. It was common in 3.5.2, which we've dropped support
for, but issue 25668 shows that it can happen in Eclipse 3.7.1 too
though it's rare.

This changeset attempts to address the issue in two ways:

(1) Spotfix the specific NPE

(2) Hold the readlock on the XML document while the detectors are
    using the DOM nodes (this is in case the nulls are related to
    this, which is not certain given that the issue is rare and not
    reproducible)

Change-Id: I14727531ea9e08abf45d70013248e702cf5a15eb