Fix SuppressLint annotations in outer classes
This changeset makes @SuppressLint work for classfile based detectors
when the error is found in an inner class and the SuppressLint
annotations is on an outer class.

Innerclasses are actually separate classes from their outer classes,
so they are processed separately (each .class file is read and
analyzed independently).

This changeset processes the class files in alphabetical* order such
that it can maintain a stack of outerclasses when processing a
class. The suppress lint check can then visit the outer class'
annotations to see if the error should be suppressed.  (*: The order
isn't exactly alphabetical: We want Foo$Bar.class to come after
Foo.class)

This changeset also tweaks the Add Annotation quickfix such that it
only offers per-method or per-class annotations, since class files do
not maintain annotation info for other granularities (such as on
variable declarations, so you cannot suppress classfile based issues
with annotations there.) We could make a lint check which ensures that
you don't try to put these annotations there :-)

(This is related to issue http://b.android.com/25948)

Change-Id: Ia9dbc39b1adc73a1b60e375edbf9b5618c7d2353