Never add R/R$* classes to the library jar files.
Previous change added excluding Manifest and BuildConfig but
made it so that it was only for the current app package.

The problem is that project that depends on libraries also generates
R classes for their libraries.
If the project itself is a library then the R classes from the libraries
would get in the library jar output.

If the main project had some diamond dependencies in its libraries
then the same R class from a library could be added twice to its
dex input causing an error.

Change-Id: I0ef4a3c3f84bf9099de27e275fa856b6c044978d