Make XML editors work with non-workspace files
This changeset makes the AndroidXmlEditor handle other editor inputs
than just FileEditorInput. In particular, it now handles
IURIEditorInput as well. This makes the "Go To Declaration" facility
for @android resources work properly again, and unlike before, you can
now use our editors in those files as well, so you can use Go To
Declaration to jump to other @android resources within the same
platform.

This changeset adds checks after all the getInputFile() calls on the
editor, which ensures that features which assume a file exists (such
as the quickfix and visual refactoring operations) do not do anything
in these files.

The layout editor has a lot of file requirements, so that editor is
deliberately skipped and you get a plain editor instead if you try to
open layout files outside of the workspace.

See
http://code.google.com/p/android/issues/detail?id=26164

Change-Id: I7457494ec0bfc2eb4e6eba2e059025142c21d3ce