Java project dependencies on libs are automatically added to main project.
Android Library projects can reference other Java-only project. This used
to be added to the library jar output. This created an issue if two Library
projects were referenced by the same app as the java project was
added to both library jar files.

This change makes it so that the java project output is not in the library
jar files but is instead added to the container of the main app as a
project classpath entry.

Change-Id: I9f8bbb104e29aede5393afd5ea4b8fe882145d49