Close the dependency file after reading it.
The dependency file will be eventually closed by the finalize() method
at some point, but it is not known when.  In the meantime, the Ant
recipe continues execution and may try to move or delete this file.
At that point the build may fail because the file is still open.

The solution is to explicitly close the BufferedReader opened by
DependencyGraph.readFile(), so the underlying file is closed as soon
as reading is finished.

Change-Id: If25f0d430191f4265a73a0e6adc3d81764c63758
Signed-off-by: Kaloian Doganov <doganov@projectoria.bg>