Fix issue with automatic jar dependency.
The previous commit relied on a cached version of the Android container
to add jar in to the list. This properly look for inside the libs folder
for jar file, and restrict looking at project classpathes to only
Java-only projects.

Also hooked into the file change listener of the Sdk class to
update the project container when a file is added to removed
to a libs folder.

Change-Id: Ia7d13b5544a27d0e8baab4873654a74b0a7889e6