fixes a bug that Working Set specified in "New Android Project" wizard is ignored in some cases.
In original code, specified Working Set data is set to mValues in ProjectNamePage#canFlipToNextPage(). This hack works in some cases, but not always.

canFlipToNextPage is invoked when user changes the name of project, but is not invoked when user changed the data of Working Set. So, if user changes WorkingSet as last action of this page, it is ignored.

I moved "mValues.workingSets = getWorkingSets();" from canFlipToNextPage() to getNextPage() because getNextPage() is alwasys invoked at the end of ProjectNamePage.

http://code.google.com/p/android/issues/detail?id=21508

Change-Id: I336fe449703144ac7a053cce6a946f931e97bb9f
Signed-off-by: YAMAZAKI Makoto <makoto1975@gmail.com>